requirements.txt
aiohttp==3.8.6
aiokafka==0.8.1
aiosignal==1.3.1
alembic==1.12.1
async-timeout==4.0.3
asyncpg==0.29.0
cryptography==41.0.5
fastapi==0.104.1
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
SQLAlchemy==2.0
uvicorn==0.23.2